display order 타입 수정

+++ src/main/resources/db/changelog/231016-banner.sql
@@ -4,7 +4,7 @@
 CREATE TABLE banner
 (
     id            bigserial    NOT NULL PRIMARY KEY,
-    display_order smallint     NOT NULL,
+    display_order integer      NOT NULL,
     link          varchar(255) NOT NULL,
     image_address varchar(255) NOT NULL,
     created_at    timestamp    NOT NULL DEFAULT NOW(),
